requirements.txt
alembic==1.13.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
redis==5.0.3
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.2
